Tighten up `Package` section of documentation
Overall just made some small things clearer/more obvious. Some are style decisions so feel free to remove. 

Also, not as many substantive changes as I thought would be made. This section though, as someone who just had to use it to create a package, is not necessarily as clear or helpful as it could be. But I don't know if it's because of this section (which is pretty clear, though awkward to have to flick to 19 to know what the various fields mean), or because there is a missing section along the lines of "Making your own package" which walks through it more. Maybe that sort of customization section might be useful?