Fix linefeed for comparison to files.
A few tests compares the output of a method to a file from src/test/resources. The files on Linux has single LF (\n) at the end of the line. When checked out on Windows the LF is converted to CRLF. This fails a string comparison. We remove the `\r` the are superflous in the